src/main/java/com/appsheet/sample/controller/UserController.java
package com.appsheet.sample.controller;


import com.appsheet.sample.service.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.GetMapping;

import java.io.IOException;

/**
 * @author Kansen Wu
 */
@Controller
public class UserController {

    private final UserService userService;

    @Autowired
    public UserController(UserService userService) {
        this.userService = userService;
    }

    @GetMapping("/")
    public String userList(Model model) throws IOException {
        model.addAttribute("users", this.userService.getUserOrderedList());
        return "home";
    }
}
